+++ ng-rosetta/fruit20/src/app/app.routing.ts
@@ -15,14 +15,17 @@ import { FruitDetailResolver }      from './app.routing.resolver';
 const appRoutes: Routes = [
     { path: 'home', component: HomeComponent },
     { path: 'fruit-list', component: FruitListComponent },
-    
+
     // Note that this route expects an id value after the detail route, it also specifies that a resolver be used
-    // during route activation. Note that the property name given to the resolve object becomes the property name on the 
+    // during route activation. Note that the property name given to the resolve object becomes the property name on the
     // ActivatedRoute's snapshot.data property in the consuming Component.
     { path: 'fruit-detail/:id', component: FruitDetailComponent, resolve: {fruitItem: FruitDetailResolver } },
 
     // The Fruit Vendor Module is Lazy Loaded via the loadChildren property. Note that this is dependent on the WebPack angular-router-loader plugin.
-    { path: 'fruit-vendor', loadChildren: './fruit-vendor/fruit-vendor.module#FruitVendorModule' },   
+    // NOTE: The fruit-vendor route is temporarily disabled as it is causing the AOT enging to blow during the prod build.
+    // This appears to only happen on windows, and only with this particular version.
+    // Uncomment the route below to enable this feature in dev.
+    // { path: 'fruit-vendor', loadChildren: './fruit-vendor/fruit-vendor.module#FruitVendorModule' },
 
     { path: '**', component: HomeComponent }
 ];